+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## \[1.2.0] - 2026-8-02
+
+* Added a configurable maximum number of luggage refills, with `-1` allowing unlimited refills.
+* Added individual configuration toggles for small, big, explorer, and ancient luggage.
+
 ## \[1.1.0] - 2026-8-02
 
 * Fixed luggages not refilling without mod installed